functions/labelimage/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
google-cloud-storage==1.42.3
google-cloud-workflows==1.5.0
google-cloud-vision==2.6.2
grpcio==1.42.0
